rearrange code and update tests
